pyproject.toml
# Declare the build backend (required)
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "nacl_middleware"
version = "0.0.2"
description = "aiohttp compatible pynacl middleware"
license = { text = "GNU General Public License v3 or later (GPLv3+)" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["pynacle", "middleware", "aiohttp"]
dependencies = [
    "aiohttp",
    "aiohttp_middlewares",
    "pynacl"
]
readme = "README.rst"


[project.optional-dependencies]
test = ["pytest"]

dev = [
    "docstring-gen",
    "build",
    "black",
    "flake8",
    "isort",
    "pip-tools",
    "bumpversion",
    "twine"
]

docs = [
    "sphinx",
    "sphinx_rtd_theme"
]

[tool.black]
line-length = 89
